Refactor the reinforcement learning module
