+++ ros_ws/src/autonomous/reinforcement_learning/scripts/drive_q_learning.py
@@ -1,25 +1,34 @@
 #!/usr/bin/env python
 
-from qlearning import QLearningNode
+from reinforcement_learning_node import ReinforcementLearningNode, device
 import os
 import rospy
-from parameters import MODEL_FILENAME
+from parameters import MODEL_FILENAME, NeuralQEstimator
 import torch
 
-
-class QLearningDrivingNode(QLearningNode):
+'''
+ROS node to drive the car using previously learned
+Q-Learning weights
+'''
+class QLearningDrivingNode(ReinforcementLearningNode):
     def __init__(self):
-        QLearningNode.__init__(self)
+        self.policy = NeuralQEstimator().to(device)
 
         try:
             self.policy.load()
+            self.policy.eval()
         except IOError:
             message = "Model parameters for the neural net not found. You need to train it first."
             rospy.logerr(message)
             rospy.signal_shutdown(message)
             exit(1)
+        
+        ReinforcementLearningNode.__init__(self)
 
     def on_receive_laser_scan(self, message):
+        if self.policy is None:
+            return
+
         state = self.convert_laser_message_to_tensor(message)
 
         with torch.no_grad():
